refactor(core): 重构项目目录结构，将核心模块移动到lib/core目录下
将types、symbolUtils、apiUtils等核心模块移动到lib/core目录
重构相关文件引用路径，更新导入语句
删除旧文件并添加新文件到版本控制
