+++ src/components/SymbolCard.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from 'react';
-import { SymbolData } from '@/lib/types';
-import { getSymbolClassName, applySymbolFont } from '@/lib/fontUtils';
+import { SymbolData } from '@/lib/core/types';
+import { getSymbolClassName, applySymbolFont } from '@/lib/font/fontUtils';
 
 interface SymbolCardProps {
   symbol: SymbolData | {
