+++ src/components/about/FeaturesSection.tsx
@@ -1,4 +1,4 @@
-import { aboutConfig, iconPaths } from '@/lib/aboutConfig';
+import { aboutConfig, iconPaths } from '@/lib/about/aboutConfig';
 import { Icon } from './Icon';
 
 // 产品特色组件
