+++ src/components/about/ProjectsSection.tsx
@@ -1,4 +1,4 @@
-import { aboutConfig, iconPaths } from '@/lib/aboutConfig';
+import { aboutConfig, iconPaths } from '@/lib/about/aboutConfig';
 import { Icon } from './Icon';
 
 // 其他作品组件
