+++ src/lib/about/aboutConfig.ts
