+++ src/lib/core/symbolData.ts
