+++ src/lib/core/symbolUtils.ts
